requirements.txt
pytest
pydicom
pynetdicom
python-dotenv
